cleanup
Remove the non-standard files

    src/syscall/syscall_nonhosted.go
    src/syscall/tables_nonhosted.go

And move contents to _tinygo.go files:

    src/syscall/syscall_tinygo.go
    src/syscall/env_tinygo.go
    src/syscall/exec_tinygo.go
